Separate ec2 and fargate
